Client: Refactorización del store
